api/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.0
sqlmodel==0.0.14
alembic==1.12.0
pydantic==2.4.0
pydantic-settings==2.0.0
psycopg[binary]==3.1.18 
python-dotenv==1.0.0
pytest==7.4.0
pytest-asyncio==0.21.0
httpx==0.24.0
